requirements.txt
# Телеграм бот
aiogram==3.3.0
aiohttp==3.9.1

# Работа с AI (Claude)
OPENROUTER==0.18.1

# Работа с переменными окружения
python-dotenv==1.0.0

# Работа с JSON и данными
pydantic==2.5.3

# Дополнительные утилиты
python-dateutil==2.8.2

# Для асинхронной работы
asyncio==3.4.3